Refactor abundance rollups into core Classifications
