+++ onecodex/models/analysis.py
@@ -1,3 +1,6 @@
+import warnings
+
+from onecodex.lib.enums import Metric
 from onecodex.models import OneCodexBase
 
 
@@ -23,10 +26,18 @@ class Analyses(OneCodexBase):
         else:
             raise NotImplementedError("No non-JSON result format implemented.")
 
+    @classmethod
+    def _transform_api_results(cls, results):
+        """Method to optionally transform the API results before caching.
+        """
+        return results
+
     def _results(self):
         try:
             if not getattr(self._resource, "_cached_result", None):
-                self._resource._cached_result = self._resource.results()
+                self._resource._cached_result = self._transform_api_results(
+                    self._resource.results()
+                )
             return self._resource._cached_result
         except AttributeError:
             raise NotImplementedError(".results() not implemented for this Analyses resource.")
@@ -39,6 +50,51 @@ class Alignments(Analyses):
 class Classifications(Analyses):
     _resource_path = "/api/v1/classifications"
     _cached_table = None
+
+    @classmethod
+    def _append_abundance_rollups(cls, results):
+        """Append cumulative abundances into the results returned from the API.
+        """
+        table = {t["tax_id"]: t.copy() for t in results}
+        for tax_id in table:
+            table[tax_id][Metric.AbundanceWChildren.value] = 0.0
+
+        # Older results might have a bug where the parent is missing,
+        # so we need to zero-out those results and renormalize the data.
+        renormalize = False
+        for tax_id, result in table.items():
+            table[tax_id] = result
+            if result[Metric.Abundance] is not None and result["parent_tax_id"] not in table:
+                renormalize = True
+                table[tax_id][Metric.Abundance] = None
+
+        if renormalize:
+            warnings.warn(
+                "Taxa with an abundance metric but no assigned reads have been removed. In order to avoid this, re-run your samples on the latest One Codex Database."
+            )
+            abundance_sum = sum([t.get(Metric.Abundance, 0) or 0 for tax_id, t in table.items()])
+            for tax_id, result in table.items():
+                if Metric.Abundance in result and result[Metric.Abundance] is not None:
+                    result[Metric.Abundance] = result[Metric.Abundance] / abundance_sum
+
+        # Roll-up abundances to parent taxa
+        for tax_id, result in table.items():
+            if result["parent_tax_id"] not in table:
+                continue
+
+            parent = table[result["parent_tax_id"]]
+            result[Metric.AbundanceWChildren.value] += result[Metric.Abundance] or 0.0
+
+            while parent:
+                parent[Metric.AbundanceWChildren.value] += result[Metric.Abundance] or 0.0
+                parent = table.get(parent["parent_tax_id"])
+
+        return list(table.values())
+
+    @classmethod
+    def _transform_api_results(cls, results):
+        results["table"] = cls._append_abundance_rollups(results["table"])
+        return results
 
     def results(self, json=True):
         """Return the complete results table for a classification.
