+++ onecodex/models/collection.py
@@ -257,48 +257,6 @@ class SampleCollection(ResourceList, AnalysisMixin):
 
         return self._cached["metadata"]
 
-    @staticmethod
-    def _calculate_abundance_rollups(results):
-        table = {t["tax_id"]: t.copy() for t in results}
-
-        # Older results might have a bug where the parent is missing,
-        # so we need to zero-out those results and renormalize the data.
-        renormalize = False
-        for tax_id, result in table.items():
-            table[tax_id] = result
-            if result[Metric.Abundance] is not None and result["parent_tax_id"] not in table:
-                renormalize = True
-                table[tax_id][Metric.Abundance] = None
-
-        if renormalize:
-            warnings.warn(
-                "Taxa with an abundance metric but no assigned reads have been removed. In order to avoid this, re-run your samples on the latest One Codex Database."
-            )
-            abundance_sum = sum([t.get(Metric.Abundance, 0) or 0 for tax_id, t in table.items()])
-            for tax_id, result in table.items():
-                if Metric.Abundance in result and result[Metric.Abundance] is not None:
-                    result[Metric.Abundance] = result[Metric.Abundance] / abundance_sum
-
-        # Roll-up abundances to parent taxa
-        for tax_id, result in table.items():
-            if Metric.AbundanceWChildren not in result:
-                result[Metric.AbundanceWChildren] = 0.0
-
-            if result["parent_tax_id"] not in table:
-                continue
-
-            parent = table[result["parent_tax_id"]]
-            result[Metric.AbundanceWChildren] += result[Metric.Abundance] or 0.0
-
-            while parent:
-                if Metric.AbundanceWChildren not in parent:
-                    parent[Metric.AbundanceWChildren] = 0.0
-
-                parent[Metric.AbundanceWChildren] += result[Metric.Abundance] or 0.0
-                parent = table.get(parent["parent_tax_id"])
-
-        return table
-
     def _collate_results(self, metric=None, include_host=None):
         """Transform a list of Classifications into `pd.DataFrames` of taxonomy and results data.
 
@@ -344,10 +302,8 @@ class SampleCollection(ResourceList, AnalysisMixin):
             results = c.results()
             host_tax_ids = results.get("host_tax_ids", [])
 
-            table = self._calculate_abundance_rollups(results["table"])
-
             # d contains info about a taxon in result, including name, id, counts, rank, etc.
-            for d in table.values():
+            for d in results["table"]:
                 d_tax_id = d["tax_id"]
 
                 if not include_host and d_tax_id in host_tax_ids:
